101-print_number.c file correction_2 commit

+++ 0x06-pointers_arrays_strings/101-print_number.c
@@ -21,5 +21,5 @@ void print_number(int n)
 	{
 		print_number(temp / 10);
 	}
-	_putchar((n % 10) + '0');
+	_putchar((unsigned int)n % 10 + '0');
 }
